+++ src/layman/layer/util.py
@@ -103,7 +103,7 @@ def get_complete_layer_info(workspace=None, layername=None, cached=False):
     if not any(partial_info):
         raise LaymanError(15, {'layername': layername})
 
-    file_type = partial_info['file']['file_type']
+    file_type = partial_info['_file_type']
     item_keys = get_layer_info_keys(file_type)
 
     complete_info = dict()